requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0
pillow==10.1.0
lxml==4.9.3
fake-useragent==1.4.0
selenium==4.15.2
webdriver-manager==4.0.1
